+++ packages/typegpu-react/src/core/root-context.tsx
@@ -179,7 +179,7 @@ export interface RootProps {
    */
   root?: TgpuRoot | undefined;
   /**
-   * (React Native only) When true, `useFrame` runs on the JS thread even if
+   * (React Native only) When true, `useFrame` runs on the RN thread even if
    * `react-native-worklets` is installed. Ignored on the web
    *
    * @default false
